refactor: PaymentClient 및 관련 클래스 패키지 이동
- PaymentClient, TossPaymentClient 등 관련 클래스의 패키지를 `client`에서 `config`로 이동
- 클래스 경로 변경에 따른 import 문 수정
- 관련 테스트 클래스의 패키지와 import 경로 동기화
