+++ src/main/java/roomescape/reservation/config/PaymentErrorHandler.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.client;
+package roomescape.reservation.config;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
